move storage settings to file
