+++ packages/server/routes/admin.js
@@ -121,33 +121,6 @@ const email_form = async (req) => {
 };
 
 /**
- * Storage settings form definition
- * @param {object} req request
- * @returns {Promise<Form>} form
- */
-const storage_form = async (req) => {
-  const form = await config_fields_form({
-    req,
-    field_names: [
-      "storage_s3_enabled",
-      "storage_s3_bucket",
-      "storage_s3_path_prefix",
-      "storage_s3_endpoint",
-      "storage_s3_region",
-      "storage_s3_access_key",
-      "storage_s3_access_secret",
-      "storage_s3_secure",
-    ],
-    action: "/admin/storage",
-  });
-  form.submitButtonClass = "btn-outline-primary";
-  form.submitLabel = req.__("Save");
-  form.onChange =
-    "remove_outline(this);$('#testemail').attr('href','#').removeClass('btn-primary').addClass('btn-outline-primary')";
-  return form;
-};
-
-/**
  * @name get
  * @function
  * @memberof module:routes/admin~routes/adminRouter
@@ -237,29 +210,6 @@ router.get(
 );
 
 /**
- * @name get/storage
- * @function
- * @memberof module:routes/admin~routes/adminRouter
- */
-router.get(
-  "/storage",
-  isAdmin,
-  error_catcher(async (req, res) => {
-    const form = await storage_form(req);
-    send_files_page({
-      res,
-      req,
-      active_sub: "Storage",
-      contents: {
-        type: "card",
-        title: req.__("Storage settings"),
-        contents: [renderForm(form, req.csrfToken())],
-      },
-    });
-  })
-);
-
-/**
  * @name get/send-test-email
  * @function
  * @memberof module:routes/admin~routes/adminRouter
@@ -315,36 +265,6 @@ router.post(
       await save_config_from_form(form);
       req.flash("success", req.__("Email settings updated"));
       res.redirect("/admin/email");
-    }
-  })
-);
-
-/**
- * @name post/email
- * @function
- * @memberof module:routes/admin~routes/adminRouter
- */
-router.post(
-  "/storage",
-  isAdmin,
-  error_catcher(async (req, res) => {
-    const form = await storage_form(req);
-    form.validate(req.body);
-    if (form.hasErrors) {
-      send_admin_page({
-        res,
-        req,
-        active_sub: "Storage",
-        contents: {
-          type: "card",
-          title: req.__("Storage settings"),
-          contents: [renderForm(form, req.csrfToken())],
-        },
-      });
-    } else {
-      await save_config_from_form(form);
-      req.flash("success", req.__("Storage settings updated"));
-      res.redirect("/admin/storage");
     }
   })
 );
